+++ src/main/kotlin/com/pswidersk/gradle/terraform/TerraformPlugin.kt
@@ -9,7 +9,7 @@ class TerraformPlugin : Plugin<Project> {
     override fun apply(project: Project): Unit = with(project) {
         extensions.create(TERRAFORM_PLUGIN_EXTENSION_NAME, TerraformPluginExtension::class.java, project)
         tasks.register<ListPropertiesTask>("listPluginProperties")
-        tasks.register<TerraformSetupTask>("terraformSetup")
+        tasks.register<TerraformSetupTask>(TERRAFORM_SETUP_TASK_NAME)
     }
 
 }